+++ resource/vim.sh
@@ -1,0 +1,2 @@
+# Install vim
+apt-get install -y vim