+++ stack/passenger-apache.sh
@@ -2,6 +2,8 @@
 bash resource/apache.sh
 bash resource/mysql-client.sh
 bash resource/ruby.sh
+bash resource/git.sh
+bash resource/vim.sh
 
 # Install passenger and enable passenger-apache module
 apt-get install -y build-essential libcurl4-openssl-dev libssl-dev zlib1g-dev apache2-prefork-dev libapr1-dev libaprutil1-dev