+++ stack/php-apache.sh
@@ -2,6 +2,8 @@
 bash resource/apache.sh
 bash resource/mysql-client.sh
 bash resource/php.sh
+bash resource/git.sh
+bash resource/vim.sh
 
 # Install php5 modules and bindings
 apt-get install --yes php5-mysql libapache2-mod-php5